Created independent modules for rxjava1 and rxjava2
