+++ build.gradle
@@ -3,6 +3,8 @@ version '0.2-SNAPSHOT'
 
 buildscript {
     ext.kotlin_version = '1.1.51'
+    ext.mockito_version = '2.8.47'
+    ext.mockito_kotlin_version = '1.5.0'
 
     repositories {
         mavenCentral()
