+++ rxjava1/src/test/kotlin/nz/co/kiwiandroiddev/mockito/rxjava/verification/rxjava1/SubscribedToObservableTest.kt
