+++ rxjava2/build.gradle
@@ -20,9 +20,18 @@ repositories {
 }
 
 dependencies {
+    compile project(':library')
+    compile "io.reactivex.rxjava2:rxjava:2.1.5"
     compile "org.jetbrains.kotlin:kotlin-stdlib-jre8:$kotlin_version"
     compile "org.mockito:mockito-core:$mockito_version"
     compile "com.nhaarman:mockito-kotlin-kt1.1:$mockito_kotlin_version"
 
     testCompile "junit:junit:4.12"
+}
+
+compileKotlin {
+    kotlinOptions.jvmTarget = "1.8"
+}
+compileTestKotlin {
+    kotlinOptions.jvmTarget = "1.8"
 }
