+++ rxjava2/src/test/kotlin/nz/co/kiwiandroiddev/mockito/rxjava/verification/rxjava2/SubscribedToObservableTest.kt
